enable validation in open api builder
