+++ polaris-core/src/main/java/org/apache/polaris/core/policy/validator/PolicyValidators.java
@@ -41,10 +41,6 @@ import org.slf4j.LoggerFactory;
 public class PolicyValidators {
   private static final Logger LOGGER = LoggerFactory.getLogger(PolicyValidators.class);
 
-  // A valid policy name should only consist of uppercase and lowercase letters (A-Z, a-z), digits
-  // (0-9), hyphens (-), underscores (_)
-  private static final Pattern POLICY_NAME_PATTERN = Pattern.compile("^[A-Za-z0-9\\-_]+$");
-
   /**
    * Validates the given policy.
    *
@@ -57,12 +53,6 @@ public class PolicyValidators {
 
     var type = PredefinedPolicyTypes.fromCode(policy.getPolicyTypeCode());
     Preconditions.checkArgument(type != null, "Unknown policy type: " + policy.getPolicyTypeCode());
-
-    Preconditions.checkArgument(
-        POLICY_NAME_PATTERN.matcher(policy.getName()).matches(),
-        "Invalid policy name: %s, A valid policy name should only consist of uppercase and lowercase letters (A-Z, a-z), digits"
-            + "(0-9), hyphens (-), underscores (_)",
-        policy.getName());
 
     switch (type) {
       case DATA_COMPACTION:
